requirements.txt
discord-py-interactions
discord.py
requests
bs4
datetime
pytz
